requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Processing
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
websocket-client==1.6.4
ccxt==4.1.30
aiohttp==3.9.0

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
joblib==1.3.2
ta==0.10.2

# Technical Analysis
TA-Lib==0.4.28
mplfinance==0.12.10a0

# Email & Notifications
email-validator==2.1.0

# WebSocket
websockets==12.0
python-socketio==5.10.0

# Utilities
schedule==1.2.0
APScheduler==3.10.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0